cleanup se0 pin setting
there is a circular dependency that has to be broken: the board
type wants to be set via USB, but you can't set the board type
until the USB enumerates. USB enumeration can be unreliable on
baosec because the SE0 pin isn't mapped as the board type defaults
to dabao.

The baosec SE0 pin is PF5 - this is inaccesible in dabao. Thus
the decision was mode to pull both possibilities during the boot1
initialization phase.

This could cause some trouble for OEM configurations that are relying
on these pins to be e.g. inputs or some other value, but this will
need to be documented as a hazard for OEM designers.